requirements.txt
asgiref==3.6.0
Django==4.1.6
django-taggit==3.1.0
sqlparse==0.4.3
tzdata==2022.7
